# docs/swagger.yaml
basePath: /
host: 127.0.0.1:8080
info:
  contact:
    email: grigorygavrin@gmail.com
  description: Link shortener service
  title: Link Shortener API
  version: "1.0"
paths:
  /:
    post:
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: Convert link to shorting and store in database
      tags:
      - SaveTXT
  /{id}:
    get:
      parameters:
      - description: 2dace3f162eb9f0d
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
      summary: Request to get the original link
      tags:
      - GetLink
  /api/shorten:
    post:
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: bad request
          schema:
            type: string
      summary: Convert link to shorting and store in database
      tags:
      - SaveJSON
  /api/shorten/batch:
    post:
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: bad request
          schema:
            type: string
      summary: Request to save data and return multiply
      tags:
      - SaveBatch
  /api/user/urls:
    delete:
      responses:
        "200":
          description: OK
          schema:
            items:
              type: array
            type: array
        "500":
          description: internal error
          schema:
            type: string
      summary: Delete handler with fan in channel
      tags:
      - DeleteBatch
    get:
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: internal error
          schema:
            type: string
      summary: Get all urls by user
      tags:
      - GetLinks
  /ping:
    get:
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Implement ping connection for sql database storage
      tags:
      - GetPing
swagger: "2.0"